requirements_CLI.txt
lxml~=4.9.1
requests~=2.28.1
packaging~=21.3
PyYAML~=6.0
markdown~=3.4.1
Pillow~=9.3.0
pyserial~=3.5
